Move parseMatches to ultils (same in all files)
