+++ src/filters/day/day.ts
@@ -1,7 +1,8 @@
 import { DAY, FILLER_WORD } from 'constants/patterns';
-import { trimWhiteSpaces, matchPattern, contains } from 'utils/Helper';
+import { matchPattern, contains } from 'utils/Helper';
 import { ParsedMatchSchema } from 'serina.schema';
 import { DateTime } from 'luxon';
+import { parseMatches } from 'utils';
 
 export default class Day {
     /*
@@ -9,22 +10,15 @@ export default class Day {
     */
    static parseText(text: string): ParsedMatchSchema[] {
         const pattern = `(${FILLER_WORD.DAY})?${DAY.ALL}`;
-        const matchForDates = matchPattern(text, pattern);
+        const matches = matchPattern(text, pattern);
 
-        if (!matchForDates) return null;
+        if (!matches) return null;
 
         // for each match, get the parsed cases
-        return matchForDates.map(elem => this.parseDateMatches(text, elem));
-    }
-
-    static parseDateMatches(text: string, matchedDate: string): ParsedMatchSchema {
-        const replaceMatch = text.toLowerCase().replace(matchedDate, '');
-
-        return {
-            text: trimWhiteSpaces(replaceMatch),
-            dateTime: this.convertMatchToDateObj(matchedDate),
-            matched: trimWhiteSpaces(matchedDate),
-        };
+        return matches.map(match => {
+            const dateTimeObj = this.convertMatchToDateObj(match);
+            return parseMatches(text, match, dateTimeObj);
+        });
     }
 
     static convertMatchToDateObj(matchingText: string): Date {
