+++ src/utils/index.ts
@@ -1,0 +1,5 @@
+import parseMatches from './parseMatches';
+
+export {
+    parseMatches,
+};
